final
Includes a toggle functionality for the buttons (instead of having to press the timer button for the entire interrogation). . 

The lie detector can only be used when a button (timerButton) is pressed, which also starts the timer. When timerButton is pressed again, the lie detector is "off" and the program outputs the duration of the interrogation in seconds (includes comments for outputting the time in minutes).

The subject is able to "plead the fifth" during the interrogation by pressing a button (interruptButton) which pauses the interrogation.